backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "boto3>=1.42.91",
    "fastapi>=0.136.0",
    "hf-transfer>=0.1.9",
    "httpx>=0.28.1",
    "huggingface-hub>=1.11.0",
    "modal>=1.4.2",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
